third_party/WebKit/LayoutTests/platform/chromium-mac-snowleopard/fast/js/stack-overflow-arrity-catch-expected.txt

Test that if an arrity check causes a stack overflow, the exception goes to the right catch

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gotRightCatch is true
PASS gotWrongCatch1 is false
FAIL gotWrongCatch2 should be false. Was true.
PASS successfullyParsed is true

TEST COMPLETE
